fix: remove assumption that strategy performance fee is equal to vaults (#303)
Before the code assumed that the performance fee of the strategy was equal to the performance fee from the vault, so total performance fee = 2 * vault performance fee. This led to incorrectly low apys.

However with the recent changes to all performance fees being handled by the vault, and not the strategy, this assumption is no longer valid.